rfc7915: Drop ICMPv4/v6 errors with untranslatable payloads
The original implementation tried to copy the ICMP error payload as a
byte sequence without modification, when the payload datagram could
not be translated into the other IP version.
This generated invalid ICMP errors that encapsulate IPv4 in ICMPv6 or
IPv6 in ICMPv4.

This patch changes the behavior to simply drop these untranslatable
ICMP errors, also fixing the bug that untranslatable payload in ICMPv4
triggers an exception.

Fixes: https://github.com/ruby-no-kai/rubykaigi-net/issues/164